adicionei a função de estado na aplicação para salvar o conteúdo compilado em um arquivo em disco
