+++ index.js
@@ -1,30 +1,15 @@
-const readline = require('readline-sync') 
-
 const robots = {
-	text:require('./robots/text.js')
+	input: require('./robots/input.js'),
+	text:require('./robots/text.js'),
+	state:require('./robots/state.js')
 }
 
 async function start(){
-	const content = {
-		maximumSentences: 10
-	}
+	robots.input()
+	await robots.text()
 
-	content.searchTerm = askAndReturnSearchTerm()
-	content.prefix = askAndReturnPrefix()
-
-	await robots.text(content)
-
-	function askAndReturnSearchTerm(){
-		return readline.question("Escreva um termo de pesquisa do Wikipedia: ")
-	}
-
-	function askAndReturnPrefix(){
-		const prefixes = ['Quem é','O que é','A história de']
-		const selectedPrefixIndex = readline.keyInSelect(prefixes, 'Escolha uma opção:')
-		const selectedPrefixText = prefixes[selectedPrefixIndex]
-		return selectedPrefixText
-	}
-	console.log(content.sentences)
+	const content = robots.state.load()
+	console.dir(content, {depth:null})
 }
 
-start();
+start()
