+++ robots/text.js
@@ -2,21 +2,26 @@ const algorithmia = require('algorithmia')
 const algorithmiaKey = require('../credentials/algorithmia.json').apiKey
 const sentenceBoundaryDetection = require('sbd')
 
-const watsonApiKey = require('../credentials/watson').apikey
+const watsonApiKey = require('../credentials/watson.json').apikey
 const NaturalLanguageUnderstandingV1 = require('watson-developer-cloud/natural-language-understanding/v1.js')
 
-var nlu = new NaturalLanguageUnderstandingV1({
+const nlu = new NaturalLanguageUnderstandingV1({
 	iam_apikey: watsonApiKey,
 	version: '2018-04-05',
 	url: 'https://gateway.watsonplatform.net/natural-language-understanding/api/'
 });
 
-async function robot(content){
+const state = require('./state.js')
+
+async function robot(){
+	const content = state.load();
 	await fetchContentFromWikipedia(content)
 	sanitizeContent(content)
 	breakContentSentences(content)
 	limitMaximumSentences(content)
 	await fetchKeywordsOfAllSentences(content)
+
+	state.save(content)
 
 	async function fetchContentFromWikipedia(content){
 		const algorithmiaAuthencitcated = algorithmia(algorithmiaKey)
